Normalize string formatting for I18n import
Previously, when collecting strings to send to crowdin, we would capture
strings surrounded by double quotes, as they appear in the .level files, then
manually escape certain combinations of quotes before dumping the values
into YAML. all in an effort to force the values that end up in the
various .en.yml files to appear properly-quoted.

HOWEVER! YAML doesn't actually require quotes for strings, except for in
a few situations. We can more easily avoid all this work and just let
the YAML library figure that out for itself.